requirements.txt
fastapi
uvicorn
httpx
nltk
scikit-learn
python-dotenv
aiohttp
sqlalchemy
asyncpg
alembic
psycopg2-binary
redis